requirements.txt
python-telegram-bot==20.7
openai==1.8.6
nest_asyncio==1.5.7
httpx==0.24.1
httpcore==0.17.2
python-dotenv
requests